requirements.txt
tqdm
numpy
scipy
pygam
joblib
pandas
scikit-learn

################################################################################
# Single cell analyses and velocity estimation
################################################################################
anndata
scanpy
dynamo-release


################################################################################
# Visualization
################################################################################
seaborn
matplotlib